Cleanup gonk_core
